+++ changelog.md
@@ -13,11 +13,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Display node data as a formatted, colour-coded grid (`da014eb`)
 - Add `GridLayout::Automatic` to allow nodes of arbitrary size to be placed in a view box (`b0f2adf`)
 - Allow for future `DataFormat` types (`6acc35f`)
+- Use `<title>` and ARIA label text to describe data type colours (``)
 
 ## Changed
 
 - Retain data node content in `model::node::NodeContent::Data`. Move node by rewriting `<g>` transform instead of child coordinates (`a11d969`)
-- Allow for data format to have configurable byte order (``)
+- Allow for data format to have configurable byte order (`7c2e423`)
 
 ## Fixed